requirements.txt
gunicorn==21.2.0
Flask==2.2.5
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
streamlit
mysql-connector-python==8.2.0
plotly==5.15.0
Flask-BasicAuth==0.2.0
matplotlib==3.7.1